+++ services/app-api/src/postgres/questionResponses/insertQuestion.ts
@@ -34,7 +34,11 @@ export async function insertQuestion(
                 },
             },
             include: {
-                documents: true,
+                documents: {
+                    orderBy: {
+                        createdAt: 'desc',
+                    },
+                },
             },
         })
