+++ services/app-api/src/postgres/questionResponses/questionHelpers.ts
