--- services/app-api/src/resolvers/questionAnswers/index.ts
@@ -1,2 +1,0 @@
-export { indexQuestionsResolver } from '../questionAnswers/indexQuestions.'
-export { createQuestionResolver } from '../questionAnswers/createQuestion'
