+++ services/app-api/src/resolvers/questionResponses/index.ts
@@ -1,0 +1,2 @@
+export { indexQuestionsResolver } from './/indexQuestions.'
+export { createQuestionResolver } from './/createQuestion'
