+++ services/app-web/src/constants/tealium.ts
@@ -45,7 +45,7 @@ const CONTENT_TYPE_BY_ROUTE: Record<RouteT | 'UNKNOWN_ROUTE', string> = {
     SUBMISSIONS_REVIEW_SUBMIT: 'form',
     SUBMISSIONS_SUMMARY: 'summary',
     SUBMISSIONS_REVISION: 'summary',
-    SUBMISSIONS_QUESTIONS_AND_ANSWERS: 'summary',
+    SUBMISSIONS_QUESTIONS_AND_RESPONSES: 'summary',
     SUBMISSIONS_UPLOAD_QUESTION: 'form',
     UNKNOWN_ROUTE: '404',
 }
