--- services/app-web/src/pages/QuestionsAnswers/index.ts
@@ -1,2 +1,0 @@
-export { QuestionsAnswers } from './QuestionsAnswers'
-export { UploadQuestions } from './UploadQuestions'
