+++ services/app-web/src/pages/QuestionsResponses/index.ts
@@ -1,0 +1,2 @@
+export { QuestionsResponses } from './QuestionsResponses'
+export { UploadQuestions } from './UploadQuestions'
